added more 'const' thoughout examples

+++ mainNonRT1/main.cpp
@@ -84,7 +84,7 @@ int main(int argc, char* argv[])
    simulation->reset();
 
    // delta time (20ms)
-   const double dt = static_cast<double>(1.0)/static_cast<double>(frameRate);
+   const double dt = 1.0 / static_cast<double>(frameRate);
 
    // execute simulation as fast as possible
    for(; simulation->getExecTimeSec() < 50.0; ) {
